+++ darwin/mobile_scanner/Sources/mobile_scanner/MobileScannerErrorCodes.swift
@@ -24,6 +24,8 @@ struct MobileScannerErrorCodes {
     static let GENERIC_ERROR_MESSAGE = "An unknown error occurred."
     // This message is used with the 'GENERIC_ERROR' error code.
     static let INVALID_ZOOM_SCALE_ERROR_MESSAGE = "The zoom scale should be between 0 and 1 (both inclusive)"
+    static let INVALID_FOCUS_POINT = "MOBILE_SCANNER_INVALID_FOCUS_POINT"
+    static let INVALID_FOCUS_POINT_MESSAGE = "The focus coordinates are not valid."
     static let NO_CAMERA_ERROR = "MOBILE_SCANNER_NO_CAMERA_ERROR"
     static let NO_CAMERA_ERROR_MESSAGE = "No cameras available."
     static let SET_SCALE_WHEN_STOPPED_ERROR = "MOBILE_SCANNER_SET_SCALE_WHEN_STOPPED_ERROR"