+++ example/lib/widgets/dialogs/box_fit_dialog.dart
@@ -34,19 +34,22 @@ class _BoxFitDialogState extends State<BoxFitDialog> {
       content: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          for (final fit in BoxFit.values)
-            RadioListTile<BoxFit>(
-              title: Text(fit.name),
-              value: fit,
-              groupValue: tempBoxFit,
-              onChanged: (BoxFit? value) {
-                if (value != null) {
-                  setState(() {
-                    tempBoxFit = value;
-                  });
-                }
-              },
+          RadioGroup<BoxFit>(
+            groupValue: tempBoxFit,
+            onChanged: (BoxFit? value) {
+              if (value != null) {
+                setState(() {
+                  tempBoxFit = value;
+                });
+              }
+            },
+            child: Column(
+              children: [
+                for (final fit in BoxFit.values)
+                  RadioListTile<BoxFit>(title: Text(fit.name), value: fit),
+              ],
             ),
+          ),
         ],
       ),
       actions: [