+++ lib/src/objects/barcode_capture.dart
@@ -33,6 +33,10 @@ class BarcodeCapture {
   /// input [image] and the [size].
   final Object? raw;
 
-  /// The size of the camera input [image].
+  /// The raw size of the camera input [image],
+  /// in which the [barcodes] were detected.
+  ///
+  /// For example if the camera resolution is 1920x1080 pixels,
+  /// this will be a [Size] with a width of 1920 and a height of 1080.
   final Size size;
 }