+++ test/mobile_scanner_test.dart
@@ -63,7 +63,7 @@ void main() {
   });
 
   tearDown(() {
-    barcodeStreamController.close();
+    unawaited(barcodeStreamController.close());
   });
 
   group('MobileScanner Widget Tests', () {